Update Paper.cpp
I think that having the fight function of the Paper, Rock, and Scissor classes return an int would be a good way of "temporarily" doubling or halving the strength of the object. We can return the doubled or halved value to the program calling it without actually changing the strength value of the object itself. Then in the main program we could compare the results of two calls to these functions to find a winner. I don't want to go any farther with the coding myself until we develop a solid design plan. Just wanted to get the basic functions and such taken care of.